D+ epoch/QSBR: reader states, chunked retirement, bounded collector, pressure gates
Rungs 1+2 plus C4 hard-bound: ACTIVE/QUIESCENT/OFFLINE reader protocol,
128-entry retire segments, 1024-entry/50us bounded collector, 16K soft /
32K hard watermarks with drain-owed latch at a safe append boundary,
maxmemory forced-drain API, DEBUG pin/hold/stats test hooks (gated),
deterministic ownership regressions (29 tests).

+++ src/networking.c
@@ -7217,7 +7217,7 @@ void ioThreadWriteToClient(client *c) {
      *   handoff (rare paths keep the battle-tested protocol). */
     if ((c->write_flags & WRITE_FLAGS_OWNED_LOCAL) && c->nwritten > 0 &&
         !(c->write_flags & WRITE_FLAGS_WRITE_ERROR) &&
-        (size_t)c->nwritten == c->io_last_bufpos && c->bufpos == (int)c->io_last_bufpos) {
+        (size_t)c->nwritten == c->io_last_bufpos && c->bufpos == c->io_last_bufpos) {
         int tid = getCurTid();
         dplus_thread_stats[tid].owned_writes++;
         dplus_thread_stats[tid].owned_net_bytes += c->nwritten;
